Tidy-up MockDialect.
